+++ src/lib/components/ToggleButton.svelte
@@ -1,33 +1,29 @@
 <script lang="ts">
-  import { type Snippet } from 'svelte'
+  import type { Snippet } from 'svelte'
   import { afterNavigate } from '$app/navigation'
   import {
     clickOutside as clickOutsideAction
   } from '$lib/actions/click-outside'
 
   interface Props {
-    clickOutside?: boolean
+    clickOutside?: HTMLElement
     open: boolean
-    target?: HTMLElement | undefined
     children?: Snippet
   }
 
   let {
-    clickOutside = false,
+    clickOutside = undefined,
     open = $bindable(),
-    target = undefined,
     children
   }: Props = $props()
 
-  const toggle = () => {
-    open = !open
-  }
+  const toggle = () => open = !open
 
   const close = (event: Event) => {
-    if (!clickOutside || !target) return
+    if (!open || !clickOutside) return
     const eventTarget = event.target as HTMLElement
 
-    if (target.contains(eventTarget)) return
+    if (clickOutside.contains(eventTarget)) return
     open = false
   }