+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 
+
+### Changed
+- Repurpose `ToggleButton.clickOutside` prop to accept the target element
+
+### Removed
+- Remove `ToggleButton.target` prop
+
 ## [0.12.1] - 2024-10-30
 
 ### Fixed